update KeyStoreType.java, load as PKCS#12 first
bugfix for PKCS#12 loading as JKS keystore type. 

FPortecle.openKeyStoreFile tries to load a keystore as a keystore of a given type until it succeeds without exceptions or null store.
Java 8+ succeeds in loading PKCS#12 stores as JKS but lacks all features a PKCS#12 store needs.

Trying to load as a PKCS#12 type first avoid this problem.
FPortecle.openKeyStoreFile uses the order given in this KeyStoreType enum.